loading datasets correctly

+++ lab-manager/src/main/java/com/redhat/iot/api/ControlData.java
@@ -30,7 +30,7 @@ public class ControlData  implements Serializable {
   private String type = null;
 
   @JsonProperty("value")
-  private Long value = null;
+  private Float value = null;
 
   @JsonProperty("units")
   private String units = null;
@@ -114,7 +114,7 @@ public class ControlData  implements Serializable {
     this.type = type;
   }
 
-  public ControlData value(Long value) {
+  public ControlData value(Float value) {
     this.value = value;
     return this;
   }
@@ -125,11 +125,11 @@ public class ControlData  implements Serializable {
   **/
   @ApiModelProperty(value = "")
 
-  public Long getValue() {
+  public Float getValue() {
     return value;
   }
 
-  public void setValue(Long value) {
+  public void setValue(Float value) {
     this.value = value;
   }
